Item CRUD-API und Projektstruktur-Bereinigung

+++ backend/src/main/java/com/datenbank/backend/repository/ItemRepository.java
@@ -1,9 +1,9 @@
 package com.datenbank.backend.repository;
 
-import com.datenbank.backend.model.Item;
+import com.datenbank.backend.entity.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ItemRepository extends JpaRepository<Item, Long> {
+public interface ItemRepository extends JpaRepository<Item, Integer> {
 }
